Use lists instead of sets to represent cloudscale data
An issue was discovered where only the first interface of two was listed. I'm
not sure how TypeSets in terraform do uniqueness, but they do it definitely
wrong in this case. Lists make sense anyway, because the cloudscale.ch
provider ensures uniqueness anyway. An advantage is also that the cloudscale.ch
API sorts lists already.

There's a test now that ensures that private and public interfaces appear both
in the terraform output.